requirements.txt
cvlib==0.2.7
numpy==1.23.2
tensorflow-gpu==2.10.1
opencv-python == 4.5.1